server/requirements.txt
asgiref==3.5.2
autopep8==1.7.0
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
coverage==6.5.0
Django==3.2.15
django-cors-headers==3.13.0
django-environ==0.9.0
django-jazzmin==2.5.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.1
drf-yasg==1.21.4
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.2.0
psycopg2==2.9.4
psycopg2-binary==2.9.4
pycodestyle==2.9.1
PyJWT==2.5.0
pyparsing==3.0.9
python-decouple==3.6
pytz==2022.4
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.3
toml==0.10.2
tzdata==2022.4
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.2.0
